+++ dmr/validation/controller.py
@@ -10,6 +10,27 @@ if TYPE_CHECKING:
     from dmr.controller import Controller
 
 
+def _raise_if_generator_endpoint(
+    controller: type['Controller[BaseSerializer]'],
+    method_name: str,
+    method: types.FunctionType,
+) -> None:
+    if inspect.isasyncgenfunction(method):
+        raise EndpointMetadataError(
+            f'{controller!r}.{method_name} is an async generator. '
+            'HTTP endpoints cannot use `yield` in method body. '
+            'Return an iterator object instead, '
+            'for example: `return self._events()`',
+        )
+    if inspect.isgeneratorfunction(method):
+        raise EndpointMetadataError(
+            f'{controller!r}.{method_name} is a sync generator. '
+            'HTTP endpoints cannot use `yield` in method body. '
+            'Return an iterator object instead, '
+            'for example: `return self._events()`',
+        )
+
+
 class ControllerValidator:
     """Validates that controller is created correctly."""
 
@@ -20,29 +41,27 @@ class ControllerValidator:
         controller: type['Controller[BaseSerializer]'],
     ) -> bool | None:
         """Run the validation."""
-        self._validate_async_generator_endpoints(controller)
+        self._validate_generator_endpoints(controller)
         is_async = self._validate_endpoints_color(controller)
         self._validate_error_handlers(controller, is_async=is_async)
         self._validate_meta_mixins(controller)
         self._validate_non_endpoints(controller)
         return is_async
 
-    def _validate_async_generator_endpoints(
+    def _validate_generator_endpoints(
         self,
         controller: type['Controller[BaseSerializer]'],
     ) -> None:
         for method_name in controller.allowed_http_methods:
             method = getattr(controller, method_name, None)
-            if isinstance(
-                method,
-                types.FunctionType,
-            ) and inspect.isasyncgenfunction(method):
-                raise EndpointMetadataError(
-                    f'{controller!r}.{method_name} is an async generator. '
-                    'HTTP endpoints cannot use `yield` in method body. '
-                    'Return an iterator object instead, '
-                    'for example: `return self._events()`',
-                )
+            if not isinstance(method, types.FunctionType):
+                continue
+
+            _raise_if_generator_endpoint(
+                controller=controller,
+                method_name=method_name,
+                method=method,
+            )
 
     def _validate_endpoints_color(
         self,