+++ CHANGELOG.md
@@ -29,6 +29,7 @@ of requirements for an API to count as public.
 - Added `strict` parameter to `request_renderer` and `request_parser`,
   added `@overload`s to both of these functions, #869
 - Allow individual `OpenAPI` views to skip schema validation, #867
+- Add endpoint validator to prevent async generator HTTP endpoints, #843
 
 ### Fixes